% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_npde.R
\name{npde.decorr.method}
\alias{npde.decorr.method}
\alias{decorr.chol}
\alias{decorr.inverse}
\alias{decorr.polar}
\title{Decorrelation methods in npde}
\arguments{
\item{x}{a square matrix
\describe{
 \item{cholesky}{decorrelation is performed through the Cholesky decomposition (default)}
 \item{inverse}{decorrelation is performed by inverting Vi through the \code{eigen} function}
 \item{polar}{the singular-value decomposition (\code{svd}) is used}
 }}
}
\description{
Specifies the method used to decorrelate observed and simulated data
}
\details{
More details can be found in the PDF documentation.
}
\references{
K. Brendel, E. Comets, C. Laffont, C. Laveille, and F.
Mentre. Metrics for external model evaluation with an application to the
population pharmacokinetics of gliclazide. \emph{Pharmaceutical Research},
23:2036--49, 2006.
}
\keyword{methods}
