% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainNpde.R
\name{npde.main}
\alias{npde.main}
\title{Main npde function}
\usage{
npde.main(object)
}
\arguments{
\item{object}{A NpdeObject object}
}
\value{
A NpdeObject object updated with the results
}
\description{
Main npde function, used to compute pd and npde
}
\keyword{internal}
\keyword{model}
