% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNpde-plotFunctions.R
\name{npde.plot.default}
\alias{npde.plot.default}
\alias{plot}
\alias{aux.scatter.box}
\alias{aux.scatter}
\alias{compute.bands.true}
\alias{compute.bands}
\alias{aux.npdeplot.meanprof}
\alias{aux.npdeplot.computepi}
\alias{aux.npdeplot.transform}
\alias{aux.npdeplot.plot}
\alias{aux.npdeplot.main}
\title{Diagnostic plots}
\usage{
npde.plot.default(npdeObject, ...)
}
\arguments{
\item{npdeObject}{an object returned by a call to \code{\link{npde}} or \code{\link{autonpde}}}

\item{\dots}{additional arguments to be passed on to the function, to control which metric (npde, pd, npd) is used or to override graphical parameters (see the PDF document for details, as well as \code{\link{set.plotoptions}})}
}
\description{
The default diagnostic plots produced after a call to \code{\link{npde}} or \code{\link{autonpde}} include a histogram of the distribution, a QQ-plot compared to the theoretical distribution, and scatterplots versus the independent variable and versus the population predictions from the model
}
\keyword{plot}
