% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNpde-plotFunctions.R
\name{npde.plot.vpc}
\alias{npde.plot.vpc}
\title{Visual Predictive Check (VPC)}
\usage{
npde.plot.vpc(npdeObject, npc=FALSE, ...)
}
\arguments{
\item{npdeObject}{an object returned by a call to \code{\link{npde}} or \code{\link{autonpde}}}

\item{npc}{a boolean, indicating whether to compute Numerical Predictive Checks (not yet implemented)}

\item{\dots}{additional arguments to be passed on to the function, to control which metric (npde, pd, npd) is used or to override graphical parameters (see the PDF document for details, as well as \code{\link{set.plotoptions}})}
}
\description{
Produces a VPC plot for the data using the simulated data provided. Note that non-stratified VPC are not suited to unbalanced designs
when features such as dose or covariates enter the model. We suggest using reference profiles instead to retain a VPC-like profile
while ensuring meaningful prediction intervals (Comets et al. 2013).
}
\references{
K. Brendel, E. Comets, C. Laffont, C. Laveille, and F.  Mentre. Metrics for external model evaluation with an application to the
population pharmacokinetics of gliclazide. \emph{Pharmaceutical Research},  23:2036--49, 2006.

E. Comets, THT Nguyen, F. Mentré. Additional features and graphs in the new npde library for R.
\emph{22nd PAGE meeting, Glasgow, UK}, 2013.
}
\seealso{
\code{\link{npde}}, \code{\link{autonpde}}, \code{\link{set.plotoptions}}
}
\author{
Emmanuelle Comets <emmanuelle.comets@bichat.inserm.fr>
}
\keyword{plot}
