% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R, R/NpdeData-methods.R
\docType{methods}
\name{read}
\alias{read}
\alias{read-methods}
\alias{read,NpdeData}
\alias{read,NpdeData-methods}
\alias{read,NpdeSimData-methods}
\alias{read,NpdeData-method}
\alias{read,NpdeSimData-method}
\title{Read data into structures (internal)}
\usage{
read(object, dat, detect=TRUE, verbose=FALSE, \dots)

\S4method{read}{NpdeData}(object, dat, detect = TRUE, verbose = FALSE, ...)

\S4method{read}{NpdeSimData}(object, dat, detect = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{object}{an object}

\item{dat}{a dataframe containing the data to be analysed}

\item{detect}{a boolean; if TRUE, automatic recognition of names will be attempted to detect necessary items 
for a NpdeData, these include longitudinal data structure, missing data and censoring information
for a NpdeSimData, the required columns should be called idsim, xsim, ysim, representing respectively simulated id, simulated 
predictor and simulated response}

\item{verbose}{a boolean; messages are printed if verbose is TRUE (defaults to FALSE)}

\item{\dots}{additional arguments for compatibility with generic}
}
\value{
an object of class \code{"\linkS4class{NpdeData}"} or \code{"\linkS4class{NpdeSimData}"}
}
\description{
Creates an NpdeData object containing the observed data, either from disk or from a dataframe
}
\keyword{internal}
\keyword{methods}
