\name{set.npde.plotoptions}
\alias{set.npde.plotoptions}
\alias{set.npde.plotoptions.NpdeData}
\alias{set.npde.plotoptions,NpdeData-method}
\alias{set.npde.plotoptions,NpdeObject-method}
\title{Set graph options}
\usage{
  set.npde.plotoptions(object, ...)
}
\arguments{
  \item{object}{an object of class NpdeData or NpdeObject}

  \item{\dots}{arguments to replace default arguments
  (currently ignored)}
}
\value{
  a list of options for graphs
}
\description{
  This function is used to set options for graphs
}
\details{
  See documentation for a list of available options.
}
\author{
  Emmanuelle Comets <emmanuelle.comets@bichat.inserm.fr>
}
\seealso{
  \code{\link{npde}}, \code{\link{autonpde}}
}
\keyword{plot}

