% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeData-methods.R, R/NpdeRes-methods.R,
%   R/NpdeObject-methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,NpdeData-method}
\alias{show.NpdeData}
\alias{show,NpdeSimData-method}
\alias{show.NpdeSimData}
\alias{show,NpdeRes-method}
\alias{show.NpdeRes}
\alias{show,NpdeObject-method}
\alias{show.NpdeObject}
\alias{show,Object-method}
\title{Displays npde objects}
\usage{
\S4method{show}{NpdeData}(object)

\S4method{show}{NpdeSimData}(object)

\S4method{show}{NpdeRes}(object)

\S4method{show}{NpdeObject}(object)
}
\arguments{
\item{object}{an object from the npde package (NpdeData, NpdeRes, NpdeObject)}
}
\description{
Prints the structure of objects from the package
}
\keyword{methods}
