% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeData.R
\name{showAll}
\alias{showAll}
\alias{showAll.NpdeData}
\alias{showAll,NpdeData-method}
\alias{showAll.NpdeSimData}
\alias{showAll,NpdeSimData-method}
\alias{showAll.NpdeRes}
\alias{showAll.NpdeObject}
\title{Brief summary of an object}
\usage{
\method{showAll}{NpdeData}(object)
}
\arguments{
\item{object}{an object}
}
\description{
Prints a brief summary of an object
}
\keyword{print}
