% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeRes.R
\name{[,NpdeRes-method}
\alias{[,NpdeRes-method}
\title{Get/set methods for NpdeRes object}
\usage{
\S4method{[}{NpdeRes}(x, i, j, drop)
}
\description{
Access slots of a NpdeRes using the object["slot"] format
}
\keyword{methods}
