% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npde.R
\docType{package}
\name{npde-package}
\alias{npde-package}
\title{Normalised prediction distribution errors for nonlinear mixed-effect models}
\description{
Routines to compute normalised prediction distribution errors, a metric
designed to evaluate non-linear mixed effect models such as those used in
pharmacokinetics and pharmacodynamics
}
\examples{
\dontrun{
data(theopp)
data(simtheopp)

# Calling autonpde with dataframes

x<-autonpde(theopp,simtheopp,ix="Time",iy="Conc",iid="ID",boolsave=FALSE)
print(x)

# Calling autonpde with names of files to be read from disk

write.table(theopp,"theopp.tab",quote=FALSE,row.names=FALSE)
write.table(simtheopp,"simtheopp.tab",quote=FALSE,row.names=FALSE)
x<-autonpde(namobs="theopp.tab", namsim="simtheopp.tab", iid = 1,
ix = 3, iy = 4, boolsave = FALSE)

print(x)
}
}
\references{
K. Brendel, E. Comets, C. Laffont, C. Laveille, and F.
Mentr{\'e}. Metrics for external model evaluation with an application to the
population pharmacokinetics of gliclazide. \emph{Pharmaceutical Research},
23:2036--49, 2006.
}
\author{
Emmanuelle Comets, Karl Brendel, Thi Huyen Tram Nguyen, France Mentre

Maintainer: Emmanuelle Comets <emmanuelle.comets@bichat.inserm.fr>
}
\keyword{models}
