% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npde-package.R
\docType{package}
\name{npde-package}
\alias{npde-package}
\alias{npde}
\title{Compute the npde.}
\description{
This package contains functions to compute an npde.
}
\author{
Michael Neely MD
\url{http://www.lapk.org}
}
\keyword{package}
