% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNpde-plotFunctions.R
\name{npde.plot.dist}
\alias{npde.plot.dist}
\title{Plots of pd/npde versus their theoretical distribution}
\usage{
npde.plot.dist(npdeObject,which="npde",dist.type="qqplot",
covsplit=FALSE,...)
}
\arguments{
\item{npdeObject}{an object returned by a call to \code{\link{npde}} or \code{\link{autonpde}}}

\item{which}{a string determining which metric to plot, one of "npde", "pd" or "npd" (defaults to "npde")}

\item{dist.type}{string, one of "ecdf" (empirical cumulative density function), "hist" (histogram) or "qqplot" (QQ-plot of the empirical distribution versus the theoretical quantiles) to determine which type of plot (default is "qqplot")}

\item{\dots}{additional arguments to be passed on to the function, to control which metric (npde, pd, npd) is used or to override graphical parameters (see the PDF document for details, as well as \code{\link{set.plotoptions}})}

\item{covsplit}{boolean. If TRUE and covariates are present in the dataset, the plots will be stratified for each covariate}
}
\description{
Produces a plot of the corresponding metric versus their theoretical distribution as a histogram,a QQ-plot, or the empirical cdf
}
\references{
K. Brendel, E. Comets, C. Laffont, C. Laveille, and F.
Mentre. Metrics for external model evaluation with an application to the
population pharmacokinetics of gliclazide. \emph{Pharmaceutical Research},
23:2036--49, 2006.
}
\seealso{
\code{\link{npde}}, \code{\link{autonpde}}, \code{\link{set.plotoptions}}
}
\author{
Emmanuelle Comets <emmanuelle.comets@bichat.inserm.fr>
}
\keyword{plot}
