% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpdeData-methods.R
\name{npdeSimData}
\alias{npdeSimData}
\title{Creates a NpdeSimData object}
\usage{
npdeSimData(npde.data, name.simdata, header=TRUE, verbose=FALSE)
}
\arguments{
\item{npde.data}{a NpdeData object}

\item{name.simdata}{name of the file containing the simulated data, or a dataframe containing it}

\item{header}{boolean indicating whether the file has a header (mandatory if
detect is TRUE)}

\item{verbose}{whether to print warning messages, defaults to FALSE (set to TRUE to check how data is being handled)}
}
\value{
an object of class NpdeSimData
}
\description{
This function is used to create a NpdeSimData object containing the simulated data corresponding to an NpdeData object
}
\seealso{
\code{\link{NpdeData}}, \code{\link{npde}}, \code{\link{autonpde}}
}
\author{
Emmanuelle Comets <emmanuelle.comets@bichat.inserm.fr>
}
