\name{showall}
\alias{showall}
\alias{showall.NpdeData}
\alias{showall,NpdeData-method}
\alias{showall.NpdeObject}
\alias{showall.NpdeRes}
\alias{showall.NpdeSimData}
\alias{showall,NpdeSimData-method}
\title{Brief summary of an object}
\usage{
  showall.NpdeData(object)
}
\arguments{
  \item{object}{an object}
}
\description{
  Prints a brief summary of an object
}
\keyword{print}

